Don't set the locale
I've removed this because on the face of it, it fails, and I'm not sure that
the base Invenio image should have opinions that might affect the locale of
Invenio deployments.

Having looked back through the history of the repository, this locale-setting
has been here since the beginning, and so there's no rationale available.