Re-pin pipenv to '<2020.5.28'
pipenv currently emits a confusing error message when used with '--system', so
pin it to a version where this wasn't happening. A fix has been committed
upstream in pypa/pipenv@cce026e but hasn't been released yet.

See #28 for the issue tracking this problem, including details of the error
message. We'll need to unpin when a new release is available.

This even gets the version string right, unlike last time ;-).